Split helpers.ts in modular files under utils/
